Algorithms
All of the images which will be used for the training stage. There is the SURF and SIFT algorithms too, where SURF algorithm will be the principal program. There is a test with only one image, and there is a test with all the images and those information will be put on a text file.